Microvias always use separate uvia netclass properties.
There is no current microvia size/drill.

Fixes https://gitlab.com/kicad/code/kicad/issues/13176

+++ pcbnew/dialogs/dialog_global_edit_tracks_and_vias.cpp
@@ -254,7 +254,7 @@ void DIALOG_GLOBAL_EDIT_TRACKS_AND_VIAS::buildNetclassesGrid()
     m_netclassGrid->AppendRows( 1 );
     buildRow( row++, settings->m_DefaultNetClass );
 
-    m_netclassGrid->AppendRows( settings->m_NetClasses.size() );
+    m_netclassGrid->AppendRows( (int) settings->m_NetClasses.size() );
 
     for( const auto& [ name, netclass ] : settings->m_NetClasses )
         buildRow( row++, netclass );
